+++ tests/parser/lexeme.cc
@@ -9,8 +9,8 @@ SCENARIO("Create some lexemes and check their values") {
     Lexeme lexeme(LexemeType::IDENTIFIER, "identifier", 13, 12);
     REQUIRE(lexeme.getType() == LexemeType::IDENTIFIER);
     REQUIRE(lexeme.getString() == "identifier");
-    REQUIRE(lexeme.getStartLineNo() == 13);
-    REQUIRE(lexeme.getStartCharNo() == 12);
+    REQUIRE(lexeme.getStartLineNo() == (unsigned long long int)13);
+    REQUIRE(lexeme.getStartCharNo() == (unsigned long long int)12);
   }
 
   /// Check that a string lexeme works as expected
@@ -18,8 +18,8 @@ SCENARIO("Create some lexemes and check their values") {
     Lexeme lexeme(LexemeType::STRING, "this is a string", 100, 1);
     REQUIRE(lexeme.getType() == LexemeType::STRING);
     REQUIRE(lexeme.getString() == "this is a string");
-    REQUIRE(lexeme.getStartLineNo() == 100);
-    REQUIRE(lexeme.getStartCharNo() == 1);
+    REQUIRE(lexeme.getStartLineNo() == (unsigned long long int)100);
+    REQUIRE(lexeme.getStartCharNo() == (unsigned long long int)1);
   }
 
   /// Check that a singularity lexeme works as expected
@@ -27,8 +27,8 @@ SCENARIO("Create some lexemes and check their values") {
     Lexeme lexeme(LexemeType::SINGULARITY, ":", 66543, 456);
     REQUIRE(lexeme.getType() == LexemeType::SINGULARITY);
     REQUIRE(lexeme.getString() == ":");
-    REQUIRE(lexeme.getStartLineNo() == 66543);
-    REQUIRE(lexeme.getStartCharNo() == 456);
+    REQUIRE(lexeme.getStartLineNo() == (unsigned long long int)66543);
+    REQUIRE(lexeme.getStartCharNo() == (unsigned long long int)456);
   }
 
   /// Check that a whitespace lexeme works as expected
